Remove InputSearch rule "more 4 characters"
In InputSeach component there was a rule that called action passed by prop when
input value was more than 4 characters.

This rule has be removed because in cases that query has less of
5 characters, the action would not be called.

One exemple is when the user to research the artist 'Xuxa'. The request
would not be made, because input value has 4 characters.